drivers: dect_nrp: upmerge/rebase changes

+++ drivers/dect_nrp/nrf91_mac/dect_nrf91_sink.c
@@ -49,7 +49,7 @@ static struct k_work_delayable lte_ipv6_router_nbr_deleted_work;
 static struct net_mgmt_event_callback dect_nrf91_net_mgmt_ipv6_event_cb;
 
 static void dect_nrf91_net_mgmt_ipv6_event_handler(struct net_mgmt_event_callback *cb,
-						   uint32_t mgmt_event, struct net_if *iface)
+						   uint64_t mgmt_event, struct net_if *iface)
 {
 	char ipv6_addr_str[NET_IPV6_ADDR_LEN];
 
@@ -278,7 +278,7 @@ static void dect_nrf91_net_mgmt_ipv6_event_handler(struct net_mgmt_event_callbac
 		break;
 
 	default:
-		LOG_WRN("Unknown event %u", mgmt_event);
+		LOG_WRN("Unknown event %llu", mgmt_event);
 		break;
 	}
 }
@@ -300,7 +300,7 @@ bool dect_nrf91_sink_ipv6_prefix_get(struct dect_nrf91_ipv6_prefix *prefix_out)
 static struct net_mgmt_event_callback net_if_cb;
 
 static void dect_nrf91_sink_net_if_mgmt_event_handler(struct net_mgmt_event_callback *cb,
-						      uint32_t event, struct net_if *iface)
+						      uint64_t event, struct net_if *iface)
 {
 	if (iface != iface_for_prefix) {
 		return;
